refactor(routes): extract playback action routes
